lud: slight clean-up
Former-commit-id: 0f21e176c29199f3e2f1f562c69d7c6b1ead0e65

+++ julia_cuda/lud/lud_kernel.jl
@@ -149,7 +149,7 @@ end
 
     sync_threads()
 
-    sum::Float32 = 0.0
+    sum = 0f0
     for i = 0:BLOCK_SIZE-1
         sum += peri_col[idx(ty, i)] * peri_row[idx(i, tx)]
     end
